+++ src/main/kotlin/dev/ekvedaras/laravelquery/inspection/UnknownColumnInspection.kt
@@ -14,11 +14,11 @@ import dev.ekvedaras.laravelquery.MyBundle
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.canHaveColumnsInArrayValues
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBlueprintMethod
-import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderClassMethod
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderMethodForColumns
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isColumnIn
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsidePhpArrayOrValue
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsideRegularFunction
+import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInteresting
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isOperatorParam
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isSchemaBuilderMethod
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.selectsAllColumns
@@ -115,18 +115,21 @@ class UnknownColumnInspection : PhpInspection() {
                 project: Project,
                 method: MethodReference,
                 expression: StringLiteralExpression
-            ) =
-                !ApplicationManager.getApplication().isReadAccessAllowed ||
+            ): Boolean {
+                val allowArray = method.name?.startsWith("where") ?: false
+
+                return !ApplicationManager.getApplication().isReadAccessAllowed ||
                     expression.containsVariable() ||
                     expression.selectsAllColumns() ||
-                    expression.isOperatorParam() ||
+                    expression.isOperatorParam(allowArray) ||
                     !method.isBuilderMethodForColumns() ||
-                    !expression.isColumnIn(method) ||
+                    !expression.isColumnIn(method, allowArray) ||
                     expression.isInsideRegularFunction() ||
                     (expression.isInsidePhpArrayOrValue() && !method.canHaveColumnsInArrayValues()) ||
-                    !method.isBuilderClassMethod(project) ||
+                    !method.isInteresting(project) ||
                     method.isSchemaBuilderMethod(project) ||
                     method.isBlueprintMethod(project)
+            }
         }
     }
 }