+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ColumnReferenceProvider.kt
@@ -11,11 +11,11 @@ import com.jetbrains.rd.util.addUnique
 import com.jetbrains.rd.util.lifetime.Lifetime
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.canHaveColumnsInArrayValues
-import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderClassMethod
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderMethodForColumns
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isColumnIn
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsidePhpArrayOrValue
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsideRegularFunction
+import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInteresting
 import dev.ekvedaras.laravelquery.utils.MethodUtils
 import dev.ekvedaras.laravelquery.utils.PsiUtils.Companion.containsVariable
 
@@ -72,12 +72,15 @@ class ColumnReferenceProvider : PsiReferenceProvider() {
         return references
     }
 
-    private fun shouldNotInspect(project: Project, method: MethodReference, element: PsiElement) =
-        !ApplicationManager.getApplication().isReadAccessAllowed ||
+    private fun shouldNotInspect(project: Project, method: MethodReference, element: PsiElement): Boolean {
+        val allowArray = method.name?.startsWith("where") ?: false
+
+        return !ApplicationManager.getApplication().isReadAccessAllowed ||
             element.containsVariable() ||
             !method.isBuilderMethodForColumns() ||
-            !element.isColumnIn(method) ||
+            !element.isColumnIn(method, allowArray) ||
             element.isInsideRegularFunction() ||
             (element.isInsidePhpArrayOrValue() && !method.canHaveColumnsInArrayValues()) ||
-            !method.isBuilderClassMethod(project)
+            !method.isInteresting(project)
+    }
 }