+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/TableOrViewReferenceProvider.kt
@@ -10,9 +10,9 @@ import com.jetbrains.php.lang.psi.elements.MethodReference
 import com.jetbrains.rd.util.addUnique
 import com.jetbrains.rd.util.lifetime.Lifetime
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
-import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderClassMethod
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderMethodByName
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsideRegularFunction
+import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInteresting
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isTableParam
 import dev.ekvedaras.laravelquery.utils.MethodUtils
 
@@ -51,5 +51,5 @@ class TableOrViewReferenceProvider : PsiReferenceProvider() {
             !method.isBuilderMethodByName() ||
             !element.isTableParam() ||
             element.isInsideRegularFunction() ||
-            !method.isBuilderClassMethod(project)
+            !method.isInteresting(project)
 }