+++ src/main/kotlin/dev/ekvedaras/laravelquery/utils/DbReferenceResolver.kt
@@ -1,11 +1,17 @@
 package dev.ekvedaras.laravelquery.utils
 
 import com.intellij.database.model.DasColumn
+import com.intellij.database.model.DasForeignKey
+import com.intellij.database.model.DasIndex
 import com.intellij.database.model.DasNamespace
 import com.intellij.database.model.DasTable
+import com.intellij.database.model.DasTableKey
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
 import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.columnsInParallel
 import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.dbDataSourcesInParallel
+import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.foreignKeysInParallel
+import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.indexesInParallel
+import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.keysInParallel
 import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.schemasInParallel
 import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.tablesInParallel
 import java.util.Collections
@@ -15,12 +21,21 @@ class DbReferenceResolver(private val reference: DbReferenceExpression) {
         val schemas = Collections.synchronizedList(reference.schema)
         val tables = Collections.synchronizedList(reference.table)
         val columns = Collections.synchronizedList(reference.column)
+        val indexes = Collections.synchronizedList(reference.index)
+        val keys = Collections.synchronizedList(reference.key)
+        val foreignKeys = Collections.synchronizedList(reference.foreignKey)
 
         when (reference.type) {
             DbReferenceExpression.Companion.Type.Table ->
                 ResolverForTableMethods(reference, schemas, tables).resolve()
             DbReferenceExpression.Companion.Type.Column ->
                 ResolverForColumnMethods(reference, schemas, tables, columns).resolve()
+            DbReferenceExpression.Companion.Type.Index ->
+                ResolverForIndexMethods(reference, indexes).resolve()
+            DbReferenceExpression.Companion.Type.Key ->
+                ResolverForKeyMethods(reference, keys).resolve()
+            DbReferenceExpression.Companion.Type.ForeignKey ->
+                ResolverForForeignKeyMethods(reference, foreignKeys).resolve()
         }
     }
 }
@@ -192,6 +207,63 @@ private class ResolverForColumnMethods(
             table.columnsInParallel()
                 .filter { it.name == reference.parts.last() }
                 .forEach { columns.add(it) }
+        }
+    }
+}
+
+private class ResolverForIndexMethods(
+    private val reference: DbReferenceExpression,
+    private val indexes: MutableList<DasIndex>,
+) {
+    fun resolve() {
+        reference.project.dbDataSourcesInParallel().forEach { dataSource ->
+            dataSource.tablesInParallel().filter {
+                reference.tablesAndAliases.containsKey(it.name)
+            }.filter {
+                reference.tablesAndAliases[it.name]?.second ?: it.dasParent?.name == it.dasParent?.name
+            }.forEach { table ->
+                table.indexesInParallel()
+                    .filter { it.name == reference.parts[0] }
+                    .forEach { indexes.add(it) }
+            }
+        }
+    }
+}
+
+private class ResolverForKeyMethods(
+    private val reference: DbReferenceExpression,
+    private val keys: MutableList<DasTableKey>,
+) {
+    fun resolve() {
+        reference.project.dbDataSourcesInParallel().forEach { dataSource ->
+            dataSource.tablesInParallel().filter {
+                reference.tablesAndAliases.containsKey(it.name)
+            }.filter {
+                reference.tablesAndAliases[it.name]?.second ?: it.dasParent?.name == it.dasParent?.name
+            }.forEach { table ->
+                table.keysInParallel()
+                    .filter { it.name == reference.parts[0] }
+                    .forEach { keys.add(it) }
+            }
+        }
+    }
+}
+
+private class ResolverForForeignKeyMethods(
+    private val reference: DbReferenceExpression,
+    private val foreignKeys: MutableList<DasForeignKey>,
+) {
+    fun resolve() {
+        reference.project.dbDataSourcesInParallel().forEach { dataSource ->
+            dataSource.tablesInParallel().filter {
+                reference.tablesAndAliases.containsKey(it.name)
+            }.filter {
+                reference.tablesAndAliases[it.name]?.second ?: it.dasParent?.name == it.dasParent?.name
+            }.forEach { table ->
+                table.foreignKeysInParallel()
+                    .filter { it.name == reference.parts[0] }
+                    .forEach { foreignKeys.add(it) }
+            }
         }
     }
 }