+++ src/test/resources/schema/dropPrimary.php
@@ -1,0 +1,3 @@
+<?php Schema::table('users', function (\Illuminate\Database\Schema\Blueprint $table) {
+    $table->dropPrimary('<caret>');
+});