+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 # Laravel Query Changelog
 
 ## [Unreleased]
+### Added
+- Index completion and reference resolving in migration files
+
+### Changed
+- Upgraded to latest gradle plugin
+- Upgraded  platform version to `2021.1.1`
+
+### Fixed
+- `create`, `update`, etc. `Eloquent` methods should only complete columns
+- Inspection errors when using nested array method in `where` methods (#35)
+- Class casting issues (#25)
+- Read access not allowed issues in `DB::table('<caret>')` (#26)
+
 ## [1.1.0]
 ### Added
 - Schema builder integration